Fix T41620
Issue here is that indices of edges will be incorrect when index of
vertices forming the edge is 3-0 or 2-0.

There are still issues here at uv edges where seam polygons intersect
each other but to solve that we need a better pass that detects uv edges
and pushes polygons along the "normal" of the edge instead of scaling
the polygon itself.